Refactor HostingEnvironment and MockFileInfo classes
Moved `HostingEnvironment` and `MockFileInfo` to the `WebOptimizer.Core.Test.Mocks` namespace.
Added XML documentation comments to `HostingEnvironment` properties and simplified its structure.
Updated `MockFileInfo` to use primary constructor syntax, removed unnecessary private fields, and streamlined property initialization and method definitions.
