+++ test/WebOptimizer.Core.Test/Mocks/MockFileInfo.cs
@@ -1,35 +1,22 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Microsoft.Extensions.FileProviders;
 
-namespace WebOptimizer.Core.Test.Mocks
+namespace WebOptimizer.Core.Test.Mocks;
+
+internal class MockFileInfo(string name, DateTimeOffset lastModified, byte[] data) : IFileInfo
 {
-    internal class MockFileInfo : IFileInfo
-    {
-        private readonly byte[] _data;
+    public bool Exists => true;
 
-        public MockFileInfo(string name, DateTimeOffset lastModified, byte[] data)
-        {
-            _data = data;
-            Name = name;
-            LastModified = lastModified;
-        }
+    public bool IsDirectory => false;
 
-        public Stream CreateReadStream()
-        {
-            return new MemoryStream(_data, false);
-        }
+    public DateTimeOffset LastModified { get; } = lastModified;
 
-        public bool Exists => true;
-        public bool IsDirectory => false;
-        public DateTimeOffset LastModified { get; }
-        public long Length => _data.Length;
-        public string Name { get; }
-        public string PhysicalPath => null;
-    }
+    public long Length => data.Length;
 
+    public string Name { get; } = name;
+
+    public string? PhysicalPath => null;
+
+    public Stream CreateReadStream() => new MemoryStream(data, false);
 }
